Return all measurements as openstruct objects
The main object will have the updatetime epoch and array of measurements.
This was done to enhance the way that we interact with the returned results
rather than working with a hash we now have objects to work with.  This is
similar to other API wrappers, such as the ruby linode implementation.