Add api calls to create and end a session
This is in preparation for the api calls to create new measurements
via the API.  This is being done by reverse engineering the scale
through network sniffing and the -debug switch with withings pairing
wizard.  Withings have also supported this effort via the forums on
a perl version of a measurement uploader.